refactor: rename app-name.ts to server-name.ts
- Rename utils/app-name.ts -> utils/server-name.ts
- Update all imports to use new filename
